+++ Conduit/CommentFeature/CommentController.cs
@@ -1,6 +1,9 @@
 ﻿using System.Security.Claims;
+using Application.Authentication;
+using Application.Authentication.Requirements;
 using Application.CommentFeature;
 using Domain.CommentFeature.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebAPI.Filters;
 
@@ -10,17 +13,19 @@ namespace WebAPI.CommentFeature;
 [ConduitExceptionHandlerFilter]
 public class CommentController : ControllerBase
 {
-    private static readonly string testUsername = "shamel";
     private readonly ICommentAppService _commentService;
-    public CommentController(ICommentAppService commentService)
+    private readonly IAuthorizationService _authorizationService;
+
+    public CommentController(ICommentAppService commentService, IAuthorizationService authorizationService)
     {
         _commentService = commentService;
+        _authorizationService = authorizationService;
     }
     
-    [HttpPost]
+    [HttpPost, Authorize]
     public async Task<IActionResult> AddComment([FromRoute] string slug, [FromBody] CommentRequest commentRequest)
     {
-        var authenticatedUsername = User.FindFirstValue("Username") ?? testUsername;
+        var authenticatedUsername = User.Identity?.Name;
         long id = await _commentService.AddCommentAsync(commentRequest, authenticatedUsername, slug);
         CommentResponse comment = await _commentService.GetCommentAsync(id);
         return Ok(new { Comment = comment });
@@ -29,16 +34,22 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetComments([FromRoute] string slug)
     {
-        var authenticatedUsername = User.FindFirstValue("Username") ?? testUsername;
+        var authenticatedUsername = User.Identity?.Name;
         IEnumerable<CommentResponse> comments = await _commentService.GetCommentsAsync(authenticatedUsername, slug);
         return Ok(new { Comments = comments });
     }
     
-    [HttpDelete("{id}")]
+    [HttpDelete("{id}"), Authorize]
     public async Task<IActionResult> DeleteComment([FromRoute] string slug, [FromRoute] long id) // todo need authorization
     {
-        var authenticatedUsername = User.FindFirstValue("Username") ?? testUsername;
-        await _commentService.DeleteCommentAsync(id);
-        return NoContent();
+        var authorizationResult =
+            await _authorizationService.AuthorizeAsync(User, new CommentId(id), CrudRequirements.Delete);
+        if (authorizationResult.Succeeded)
+        {
+            await _commentService.DeleteCommentAsync(id);
+            return NoContent();
+        }
+        
+        return Forbid();
     }
 }
